fix: Nested catch and create domain module
