+++ ItemDomain/ItemDomain.h
@@ -1,0 +1,18 @@
+//
+//  ItemDomain.h
+//  ItemDomain
+//
+//  Created by Ahmed on 11/05/2024.
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for ItemDomain.
+FOUNDATION_EXPORT double ItemDomainVersionNumber;
+
+//! Project version string for ItemDomain.
+FOUNDATION_EXPORT const unsigned char ItemDomainVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <ItemDomain/PublicHeader.h>
+
+
