+++ RetryPoc.xcodeproj/xcuserdata/ahamed.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,6 +4,11 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
+		<key>ItemDomain.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>1</integer>
+		</dict>
 		<key>RetryPoc.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
